Add code format suggestions from pr and ci

+++ src/base/bittorrent/torrentimpl.cpp
@@ -1200,7 +1200,7 @@ qlonglong TorrentImpl::eta() const
         if (maxInactiveSeedingTimeValue >= 0)
         {
             inactiveSeedingTimeEta = (maxInactiveSeedingTimeValue * 60) - timeSinceActivity();
-            inactiveSeedingTimeEta = std::max(inactiveSeedingTimeEta, 0);
+            inactiveSeedingTimeEta = std::max(inactiveSeedingTimeEta, qlonglong(0));
         }
 
         return std::min({ratioEta, seedingTimeEta, inactiveSeedingTimeEta});
